Move attribute logic to module and add tests
