+++ cache_diff_derive/src/fields.rs
@@ -1,102 +1,9 @@
+use crate::attributes::CacheAttributes;
 use proc_macro2::TokenStream;
 use quote::quote;
-use std::str::FromStr;
-use strum::IntoEnumIterator;
-use syn::punctuated::Punctuated;
 use syn::Data::Struct;
 use syn::Fields::Named;
-use syn::{Attribute, DataStruct, FieldsNamed, Ident, PathArguments, Token};
-use syn::{DeriveInput, Field};
-
-/// Holds the one or more attributes from `#[cache_diff(...)]` attribute configurationo
-///
-#[derive(Debug, PartialEq, Eq, Default)]
-struct CacheAttributes {
-    /// When present indicates the given string should be used as a name instead of the field name
-    rename: Option<String>,
-
-    /// When present indicates the given path to a function should be used to customize the display of the field value
-    display: Option<syn::Path>,
-
-    /// When `Some` indicates the field should be ignored in the diff comparison
-    ignore: Option<()>,
-}
-
-impl CacheAttributes {
-    /// Parse all attributes inside of `#[cache_diff(...)]` and return a single CacheAttributes value
-    fn parse_all(input: &Attribute) -> syn::Result<Self> {
-        let mut attribute = CacheAttributes::default();
-
-        match &input.meta {
-            syn::Meta::List(meta_list) => {
-                for attr in meta_list
-                    .parse_args_with(Punctuated::<CacheAttributes, Token![,]>::parse_terminated)?
-                {
-                    if let Some(value) = attr.rename {
-                        attribute.rename = Some(value);
-                    }
-                    if let Some(display) = attr.display {
-                        attribute.display = Some(display);
-                    }
-                    if let Some(ignore) = attr.ignore {
-                        attribute.ignore = Some(ignore);
-                    }
-                }
-                Ok(attribute)
-            }
-            _ => Err(syn::Error::new(
-                input.pound_token.span,
-                "Expected a list of attributes",
-            )),
-        }
-    }
-}
-
-/// Valid keys for the `#[cache_diff(...)]` attribute
-///
-/// Used in parsing the user input and validating it
-#[derive(Debug, strum::EnumIter, strum::EnumString, PartialEq, strum::Display)]
-#[allow(non_camel_case_types)]
-enum AttributeKey {
-    rename,  // #[cache_diff(rename="...")]
-    display, // #[cache_diff(display="...")]
-    ignore,  // #[cache_diff(ignore)]
-}
-
-impl syn::parse::Parse for CacheAttributes {
-    // Parse a single attribute inside of a `#[cache_diff(...)]` attribute
-    fn parse(input: syn::parse::ParseStream) -> syn::Result<Self> {
-        let name: Ident = input.parse()?;
-        let name_str = name.to_string();
-        let mut attribute = CacheAttributes::default();
-        match AttributeKey::from_str(&name_str).map_err(|_| {
-            syn::Error::new(
-                name.span(),
-                format!(
-                    "Unknown attribute: {name_str}. Must be one of {}",
-                    AttributeKey::iter()
-                        .map(|k| format!("`{k}`"))
-                        .collect::<Vec<String>>()
-                        .join(", ")
-                ),
-            )
-        })? {
-            AttributeKey::rename => {
-                input.parse::<syn::Token![=]>()?;
-                let value = input.parse::<syn::LitStr>()?;
-                attribute.rename = Some(value.value());
-            }
-            AttributeKey::display => {
-                input.parse::<syn::Token![=]>()?;
-                attribute.display = Some(input.parse()?);
-            }
-            AttributeKey::ignore => {
-                attribute.ignore = Some(());
-            }
-        }
-        Ok(attribute)
-    }
-}
+use syn::{DataStruct, DeriveInput, Field, FieldsNamed, PathArguments};
 
 fn extract_attribute_from_field<'a>(f: &'a Field, name: &'a str) -> Option<&'a syn::Attribute> {
     f.attrs.iter().find(|&attr| attr.path().is_ident(name))
