+++ cache_diff_derive/src/lib.rs
@@ -1,6 +1,7 @@
 use fields::create_cache_diff;
 use proc_macro::TokenStream;
 
+mod attributes;
 mod fields;
 
 #[proc_macro_derive(CacheDiff, attributes(cache_diff))]
